requirements.txt
upsonic
python-dotenv
httpx